refactor(settings): split settings_window into directory module
Convert settings_window.rs (2173 LOC) into a directory module and
extract two submodules:

- queue_inspector.rs: queue inspector dialog, about dialog (185 LOC)
- watch_folders.rs: folder row construction, rules dialog, album picker (420 LOC)

mod.rs retains the main build_settings_window_with_parent function
and status polling (1586 LOC).

All re-exports preserved -- crate::settings_window::show_queue_inspector
and crate::settings_window::build_settings_window_with_parent remain
stable. All 101 tests pass, clippy clean.
