+++ src/settings_window/mod.rs
@@ -6,10 +6,7 @@ use crate::diagnostics;
 use adw::prelude::*;
 use glib::clone;
 use gtk::prelude::*;
-use gtk::{
-    Box, Button, Entry, FileDialog, ListBox, Orientation, PasswordEntry, ProgressBar,
-    ScrolledWindow, Switch,
-};
+use gtk::{Button, Entry, FileDialog, ListBox, PasswordEntry, ProgressBar, ScrolledWindow, Switch};
 use libadwaita as adw;
 use std::cell::Cell;
 use std::cell::RefCell;
@@ -20,8 +17,13 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use crate::app_context::AppContext;
-use crate::queue_manager::QueueManager;
-use crate::watch_path_display::{display_watch_path, watch_path_subtitle};
+
+mod queue_inspector;
+mod watch_folders;
+
+use queue_inspector::show_about_dialog;
+pub use queue_inspector::show_queue_inspector;
+use watch_folders::add_folder_row;
 
 /// Holds GTK widgets for a single watch-folder row in the settings list.
 struct FolderRowData {
@@ -1571,585 +1573,6 @@ pub fn build_settings_window_with_parent(
         ),
     );
     window.present();
-}
-
-fn add_folder_row(
-    list: &ListBox,
-    entry: &WatchPathEntry,
-    albums_ref: Rc<RefCell<Vec<(String, String)>>>,
-    tracked_rows: &Rc<RefCell<Vec<FolderRowData>>>,
-    on_settings_changed: Rc<dyn Fn()>,
-) {
-    let path = entry.path().to_string();
-    let base_subtitle = watch_path_subtitle(&path).unwrap_or_default().to_string();
-
-    let mut initial_subtitle = base_subtitle.clone();
-    if !initial_subtitle.is_empty() {
-        initial_subtitle.push('\n');
-    }
-    initial_subtitle.push_str("Status: Idle");
-
-    let expander_row = adw::ExpanderRow::builder()
-        .title(display_watch_path(&path))
-        .subtitle(&initial_subtitle)
-        .subtitle_lines(2)
-        .title_lines(1)
-        .build();
-
-    let album_name = Rc::new(RefCell::new(
-        entry
-            .album_name()
-            .unwrap_or(DEFAULT_ALBUM_LABEL)
-            .to_string(),
-    ));
-
-    let rules = Rc::new(RefCell::new(entry.rules()));
-
-    let picker_btn = Button::builder()
-        .label(format!("Album: {}", album_name.borrow()))
-        .valign(gtk::Align::Center)
-        .tooltip_text("Select or create a target Immich album")
-        .build();
-
-    let picker_btn_clone = picker_btn.clone();
-    let album_name_clone = album_name.clone();
-    let albums_ref_clone = albums_ref.clone();
-    let on_settings_changed_for_picker = on_settings_changed.clone();
-
-    picker_btn.connect_clicked(clone!(
-        #[weak]
-        expander_row,
-        move |_| {
-            if let Some(window) = expander_row
-                .root()
-                .and_then(|root| root.downcast::<gtk::Window>().ok())
-            {
-                let window_clone = window.clone();
-                let albums_ref_clone = albums_ref_clone.clone();
-                let album_name_clone = album_name_clone.clone();
-                let picker_btn_clone = picker_btn_clone.clone();
-                let on_settings_changed_for_picker = on_settings_changed_for_picker.clone();
-
-                glib::idle_add_local_once(move || {
-                    show_album_picker_dialog(
-                        &window_clone,
-                        albums_ref_clone,
-                        album_name_clone,
-                        picker_btn_clone,
-                        on_settings_changed_for_picker,
-                    );
-                });
-            }
-        }
-    ));
-
-    let album_subrow = adw::ActionRow::builder().title("Target Album").build();
-    album_subrow.add_suffix(&picker_btn);
-    expander_row.add_row(&album_subrow);
-
-    let remove_btn = Button::builder()
-        .icon_name("user-trash-symbolic")
-        .valign(gtk::Align::Center)
-        .css_classes(vec!["destructive-action".to_string()])
-        .build();
-    let rules_btn = Button::builder()
-        .label("Rules")
-        .tooltip_text("Edit folder rules")
-        .valign(gtk::Align::Center)
-        .build();
-
-    let list_clone = list.clone();
-    let tracked_clone = tracked_rows.clone();
-    let path_clone = path.clone();
-    let rules_clone = rules.clone();
-    let path_for_rules = path.clone();
-    let on_settings_changed_for_rules = on_settings_changed.clone();
-
-    rules_btn.connect_clicked(clone!(
-        #[weak]
-        expander_row,
-        move |_| {
-            if let Some(window) = expander_row
-                .root()
-                .and_then(|root| root.downcast::<gtk::Window>().ok())
-            {
-                let window = window.clone();
-                let path_for_rules = path_for_rules.clone();
-                let rules_clone = rules_clone.clone();
-                let on_settings_changed_for_rules = on_settings_changed_for_rules.clone();
-                glib::idle_add_local_once(move || {
-                    show_folder_rules_dialog(
-                        &window,
-                        &path_for_rules,
-                        rules_clone,
-                        on_settings_changed_for_rules,
-                    );
-                });
-            }
-        }
-    ));
-
-    let on_settings_changed_for_remove = on_settings_changed.clone();
-
-    remove_btn.connect_clicked(clone!(
-        #[weak]
-        expander_row,
-        move |_| {
-            let list_clone = list_clone.clone();
-            let tracked_clone = tracked_clone.clone();
-            let path_clone = path_clone.clone();
-            let expander_row = expander_row.clone();
-            let on_settings_changed_for_remove = on_settings_changed_for_remove.clone();
-            glib::idle_add_local_once(move || {
-                if let Some(focus_target) = list_clone.first_child() {
-                    focus_target.grab_focus();
-                }
-                list_clone.remove(&expander_row);
-                tracked_clone.borrow_mut().retain(|r| r.path != path_clone);
-                (on_settings_changed_for_remove)();
-            });
-        }
-    ));
-
-    let rules_subrow = adw::ActionRow::builder().title("Folder Rules").build();
-    rules_subrow.add_suffix(&rules_btn);
-    expander_row.add_row(&rules_subrow);
-
-    let remove_subrow = adw::ActionRow::builder().title("Remove Folder").build();
-    remove_subrow.add_suffix(&remove_btn);
-    expander_row.add_row(&remove_subrow);
-
-    list.append(&expander_row);
-    tracked_rows.borrow_mut().push(FolderRowData {
-        path,
-        album_name,
-        rules,
-        action_row: expander_row,
-        base_subtitle,
-    });
-}
-
-fn show_folder_rules_dialog(
-    parent: &impl gtk::prelude::IsA<gtk::Window>,
-    folder_path: &str,
-    rules_state: Rc<RefCell<FolderRules>>,
-    on_settings_changed: Rc<dyn Fn()>,
-) {
-    let dialog = adw::Window::builder()
-        .transient_for(parent)
-        .modal(true)
-        .title("Folder Rules")
-        .default_width(420)
-        .build();
-    let content = Box::builder()
-        .orientation(Orientation::Vertical)
-        .spacing(12)
-        .margin_top(12)
-        .margin_bottom(12)
-        .margin_start(12)
-        .margin_end(12)
-        .build();
-    dialog.set_content(Some(&content));
-
-    let title = gtk::Label::builder()
-        .label(format!("Rules for {}", display_watch_path(folder_path)))
-        .halign(gtk::Align::Start)
-        .wrap(true)
-        .build();
-    content.append(&title);
-
-    let current = rules_state.borrow().clone();
-
-    let list_box = gtk::ListBox::builder()
-        .selection_mode(gtk::SelectionMode::None)
-        .css_classes(vec![String::from("boxed-list")])
-        .build();
-
-    let ignore_hidden = adw::SwitchRow::builder()
-        .title("Ignore Hidden Files / Folders")
-        .subtitle("Skip paths that contain hidden components such as .cache or .thumbnails.")
-        .active(current.ignore_hidden)
-        .build();
-
-    list_box.append(&ignore_hidden);
-    content.append(&list_box);
-
-    let max_size_entry = Entry::builder()
-        .placeholder_text("Max file size in MB, leave blank for no limit")
-        .width_request(0)
-        .text(
-            current
-                .max_file_size_mb
-                .map(|value| value.to_string())
-                .unwrap_or_default(),
-        )
-        .build();
-    content.append(&max_size_entry);
-
-    let extensions_entry = Entry::builder()
-        .placeholder_text("Comma-separated extensions, e.g. jpg,png,mp4")
-        .width_request(0)
-        .text(current.allowed_extensions.join(", "))
-        .build();
-    content.append(&extensions_entry);
-
-    let actions = Box::builder()
-        .orientation(Orientation::Horizontal)
-        .spacing(8)
-        .halign(gtk::Align::End)
-        .build();
-    let cancel_btn = Button::builder().label("Cancel").build();
-    let save_btn = Button::builder()
-        .label("Save")
-        .css_classes(vec!["suggested-action".to_string()])
-        .build();
-    actions.append(&cancel_btn);
-    actions.append(&save_btn);
-    content.append(&actions);
-
-    cancel_btn.connect_clicked(clone!(
-        #[weak]
-        dialog,
-        move |_| {
-            dialog.close();
-        }
-    ));
-
-    save_btn.connect_clicked(clone!(
-        #[weak]
-        dialog,
-        move |_| {
-            let max_file_size_mb = max_size_entry.text().trim().parse::<u64>().ok();
-            let allowed_extensions = extensions_entry
-                .text()
-                .split(',')
-                .map(|part| part.trim().trim_start_matches('.').to_ascii_lowercase())
-                .filter(|part| !part.is_empty())
-                .collect::<Vec<_>>();
-
-            *rules_state.borrow_mut() = FolderRules {
-                ignore_hidden: ignore_hidden.is_active(),
-                max_file_size_mb,
-                allowed_extensions,
-            };
-            (on_settings_changed)();
-            dialog.close();
-        }
-    ));
-
-    dialog.present();
-}
-
-pub fn show_queue_inspector(
-    parent: &impl gtk::prelude::IsA<gtk::Window>,
-    queue_manager: Arc<QueueManager>,
-) {
-    let dialog = adw::Window::builder()
-        .transient_for(parent)
-        .modal(true)
-        .title("Queue Inspector")
-        .default_width(900)
-        .default_height(680)
-        .build();
-    let content = Box::builder()
-        .orientation(Orientation::Vertical)
-        .spacing(12)
-        .margin_top(12)
-        .margin_bottom(12)
-        .margin_start(12)
-        .margin_end(12)
-        .build();
-    dialog.set_content(Some(&content));
-
-    let actions = Box::builder()
-        .orientation(Orientation::Horizontal)
-        .spacing(8)
-        .halign(gtk::Align::End)
-        .build();
-    content.append(&actions);
-
-    let retry_all_btn = Button::builder().label("Retry All Failed").build();
-    let clear_failed_btn = Button::builder().label("Clear Failed Queue").build();
-    actions.append(&retry_all_btn);
-    actions.append(&clear_failed_btn);
-
-    let failed_group = adw::PreferencesGroup::builder()
-        .title("Failed Retry Queue")
-        .build();
-    content.append(&failed_group);
-
-    let failed_tasks = queue_manager.failed_tasks();
-    if failed_tasks.is_empty() {
-        failed_group.add(
-            &adw::ActionRow::builder()
-                .title("No failed items")
-                .subtitle("The retry queue is currently empty.")
-                .build(),
-        );
-    } else {
-        for task in failed_tasks {
-            let row = adw::ActionRow::builder()
-                .title(
-                    Path::new(&task.path)
-                        .file_name()
-                        .and_then(|name| name.to_str())
-                        .unwrap_or(task.path.as_str()),
-                )
-                .subtitle(&task.path)
-                .subtitle_lines(3)
-                .build();
-            let retry_btn = Button::builder().label("Retry").build();
-            let task_path = task.path.clone();
-            let qm = queue_manager.clone();
-            retry_btn.connect_clicked(move |btn| {
-                btn.set_sensitive(false);
-                let qm = qm.clone();
-                let task_path = task_path.clone();
-                glib::MainContext::default().spawn_local(async move {
-                    let _ = qm.retry_failed_path(&task_path).await;
-                });
-            });
-            row.add_suffix(&retry_btn);
-            failed_group.add(&row);
-        }
-    }
-
-    let events_group = adw::PreferencesGroup::builder()
-        .title("Recent Queue Activity")
-        .build();
-    content.append(&events_group);
-
-    let events_scroll = ScrolledWindow::builder()
-        .min_content_height(340)
-        .vexpand(true)
-        .build();
-    let events_list = ListBox::builder()
-        .selection_mode(gtk::SelectionMode::None)
-        .css_classes(vec!["boxed-list".to_string()])
-        .build();
-    events_scroll.set_child(Some(&events_list));
-    events_group.add(&events_scroll);
-
-    for event in queue_manager.recent_events() {
-        let row = adw::ActionRow::builder()
-            .title(
-                Path::new(&event.path)
-                    .file_name()
-                    .and_then(|name| name.to_str())
-                    .unwrap_or(event.path.as_str()),
-            )
-            .subtitle(format!(
-                "{} | attempts={}{}",
-                event.status,
-                event.attempts,
-                event
-                    .detail
-                    .as_ref()
-                    .map(|detail| format!(" | {}", detail))
-                    .unwrap_or_default()
-            ))
-            .subtitle_lines(4)
-            .build();
-        row.add_prefix(
-            &gtk::Label::builder()
-                .label(display_watch_path(&event.path))
-                .wrap(true)
-                .halign(gtk::Align::Start)
-                .build(),
-        );
-        events_list.append(&row);
-    }
-
-    let qm_retry_all = queue_manager.clone();
-    retry_all_btn.connect_clicked(move |btn| {
-        btn.set_sensitive(false);
-        let qm = qm_retry_all.clone();
-        glib::MainContext::default().spawn_local(async move {
-            let _ = qm.retry_all_failed().await;
-        });
-    });
-
-    let qm_clear = queue_manager.clone();
-    clear_failed_btn.connect_clicked(move |_| {
-        let _ = qm_clear.clear_failed();
-    });
-
-    let close_btn = Button::builder().label("Close").build();
-    close_btn.connect_clicked(clone!(
-        #[weak]
-        dialog,
-        move |_| {
-            dialog.close();
-        }
-    ));
-    content.append(&close_btn);
-    dialog.present();
-}
-
-fn show_about_dialog(parent: &impl gtk::prelude::IsA<gtk::Widget>) {
-    let about = adw::AboutDialog::builder()
-        .application_name("Mimick")
-        .application_icon("dev.nicx.mimick")
-        .version(env!("CARGO_PKG_VERSION"))
-        .developer_name("Nick Cardoso")
-        .website("https://github.com/nicx17/mimick")
-        .issue_url("https://github.com/nicx17/mimick/issues")
-        .license_type(gtk::License::Gpl30)
-        .build();
-
-    about.add_credit_section(
-        Some("Icon Design"),
-        &["Round Icons https://unsplash.com/illustrations/a-white-and-orange-flower-on-a-white-background-IkQ_WrJzZOM"],
-    );
-
-    let third_party = glib::markup_escape_text(include_str!("../THIRD_PARTY_LICENSES_SUMMARY.txt"));
-    about.add_legal_section(
-        "Third-party Licenses",
-        None,
-        gtk::License::Custom,
-        Some(&third_party),
-    );
-
-    about.present(Some(parent));
-}
-
-fn show_album_picker_dialog(
-    parent: &impl gtk::prelude::IsA<gtk::Window>,
-    albums_ref: Rc<RefCell<Vec<(String, String)>>>,
-    target_album_state: Rc<RefCell<String>>,
-    trigger_btn: Button,
-    on_settings_changed: Rc<dyn Fn()>,
-) {
-    let dialog = adw::Window::builder()
-        .transient_for(parent)
-        .modal(true)
-        .title("Select Album")
-        .default_width(400)
-        .default_height(500)
-        .build();
-
-    let header_bar = adw::HeaderBar::new();
-    let vbox = Box::builder().orientation(Orientation::Vertical).build();
-    dialog.set_content(Some(&vbox));
-    vbox.append(&header_bar);
-
-    let search_entry = gtk::SearchEntry::builder()
-        .halign(gtk::Align::Center)
-        .width_request(300)
-        .margin_top(8)
-        .margin_bottom(8)
-        .build();
-    vbox.append(&search_entry);
-
-    let list_box = gtk::ListBox::builder()
-        .selection_mode(gtk::SelectionMode::None)
-        .margin_start(12)
-        .margin_end(12)
-        .margin_bottom(12)
-        .build();
-    list_box.add_css_class("boxed-list");
-
-    let scrolled_window = gtk::ScrolledWindow::builder()
-        .hscrollbar_policy(gtk::PolicyType::Never)
-        .vscrollbar_policy(gtk::PolicyType::Automatic)
-        .vexpand(true)
-        .build();
-    scrolled_window.set_child(Some(&list_box));
-    vbox.append(&scrolled_window);
-
-    // Dynamic filtering capability
-    let albums_ref_cloned = albums_ref.clone();
-    let apply_filter = {
-        let list_box = list_box.clone();
-        let dialog = dialog.clone();
-        let target_album_state = target_album_state.clone();
-        let trigger_btn = trigger_btn.clone();
-
-        move |query: &str| {
-            // Clear existing
-            while let Some(child) = list_box.first_child() {
-                list_box.remove(&child);
-            }
-
-            let q = query.trim().to_lowercase();
-
-            // Row 1: Default Folder Name (only if it matches search)
-            if q.is_empty() || DEFAULT_ALBUM_LABEL.to_lowercase().contains(&q) {
-                let default_row = adw::ActionRow::builder()
-                    .title(DEFAULT_ALBUM_LABEL)
-                    .subtitle("Creates album dynamically per-folder")
-                    .activatable(true)
-                    .build();
-                let dialog_clone = dialog.clone();
-                let state_clone = target_album_state.clone();
-                let btn_clone = trigger_btn.clone();
-                let on_settings_changed_clone = on_settings_changed.clone();
-                default_row.connect_activated(move |_| {
-                    *state_clone.borrow_mut() = DEFAULT_ALBUM_LABEL.to_string();
-                    btn_clone.set_label(&format!("Album: {}", DEFAULT_ALBUM_LABEL));
-                    (on_settings_changed_clone)();
-                    dialog_clone.close();
-                });
-                list_box.append(&default_row);
-            }
-
-            // Row 2: Create Custom (if query is typed)
-            if !q.is_empty() {
-                let typed_raw = query.trim().to_string();
-                let create_row = adw::ActionRow::builder()
-                    .title(format!("Create new: \"{}\"", typed_raw))
-                    .activatable(true)
-                    .build();
-                let dialog_clone = dialog.clone();
-                let state_clone = target_album_state.clone();
-                let btn_clone = trigger_btn.clone();
-                let on_settings_changed_clone = on_settings_changed.clone();
-                create_row.connect_activated(move |_| {
-                    *state_clone.borrow_mut() = typed_raw.clone();
-                    btn_clone.set_label(&format!("Album: {}", typed_raw));
-                    (on_settings_changed_clone)();
-                    dialog_clone.close();
-                });
-                list_box.append(&create_row);
-            }
-
-            // Row 3+: Remote Albums
-            for (name, _) in albums_ref_cloned.borrow().iter() {
-                if name == DEFAULT_ALBUM_LABEL {
-                    continue; // Skip the "Use default folder name" if we pushed it above
-                }
-                if q.is_empty() || name.to_lowercase().contains(&q) {
-                    let album_name = name.clone();
-                    let row = adw::ActionRow::builder()
-                        .title(&album_name)
-                        .activatable(true)
-                        .build();
-                    let dialog_clone = dialog.clone();
-                    let state_clone = target_album_state.clone();
-                    let btn_clone = trigger_btn.clone();
-                    let album_name_clone = album_name.clone();
-                    let on_settings_changed_clone = on_settings_changed.clone();
-                    row.connect_activated(move |_| {
-                        *state_clone.borrow_mut() = album_name_clone.clone();
-                        btn_clone.set_label(&format!("Album: {}", album_name_clone));
-                        (on_settings_changed_clone)();
-                        dialog_clone.close();
-                    });
-                    list_box.append(&row);
-                }
-            }
-        }
-    };
-
-    // Initial populate
-    apply_filter("");
-
-    let apply_filter_rc = Rc::new(apply_filter);
-    search_entry.connect_search_changed(move |entry| {
-        apply_filter_rc(&entry.text());
-    });
-
-    dialog.present();
 }
 
 #[cfg(test)]
